feat: Add segmentation configuration to ImageLoadedEvent
This commit adds a new attribute `segmentation_configuration` to the `ImageLoadedEvent` class. The attribute is of type `dict | None`. It allows passing a segmentation configuration dictionary when creating an instance of `ImageLoadedEvent`.

The following significant changes were made:
- Added the attribute `segmentation_configuration` to the `ImageLoadedEvent` class.
- Modified the constructor of the `ImageLaodedEvent` class to accept the new attribute.
- Updated the `_load_image_node()` method in the `ImageNodeParserWorklof` class to pass the segmentation configuration when creating an instance of `ImageLaodedEvent`.
- Updated the `_parse_image_node()` method in the `ImageNodeParserWorklof` class to receive and use a segmentation configuration dictionary.

These changes enable users to provide custom segmentation configurations for image loading and parsing, allowing for more flexibility and customization.